+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:uic/list_uic.dart';
+import 'package:uic/progress_uic.dart';
 
 void main() => runApp(MyApp());
 
@@ -11,7 +12,8 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'ListUic Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        primarySwatch: Colors.blueGrey,
+        accentColor: Colors.amber,
       ),
       home: MyHomePage(title: 'ListUic Demo'),
     );
@@ -38,7 +40,8 @@ class _MyHomePageState extends State<MyHomePage> {
     _controller = ListUicController<int>(
 //      items: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10],
       onGetItems: (int page) => _getItems(page),
-      initialLoading: true,
+//      initialLoading: false,
+//      allowPagination: false,
     );
   }
 
@@ -56,19 +59,55 @@ class _MyHomePageState extends State<MyHomePage> {
             subtitle: Text('Subtitle ${item}'),
           );
         },
+//        emptyDataIcon: Icon(Icons.refresh, size: 72.0, color: Colors.amberAccent),
+//        emptyDataText: "You don't have any item",
+//        emptyDataView: Center(
+//          child: Text("Empty",
+//            style: Theme.of(context).textTheme.headline1,
+//          ),
+//        ),
+//        emptyErrorIcon: Icon(Icons.error, size: 72.0, color: Colors.redAccent),
+//        emptyErrorText: "Data loading failed",
+//        emptyErrorView: Center(
+//          child: Text("Failed",
+//            style: Theme.of(context).textTheme.headline1,
+//          ),
+//        ),
+//        emptyProgressText: "Please wait...",
+//        emptyProgressView: Center(
+//          child: Text("Wait...",
+//            style: Theme.of(context).textTheme.headline1,
+//          ),
+//        ),
+//        errorText: "Something went wrong",
+//        errorColor: Colors.orangeAccent,
+//        nextPageProgressView: ProgressUic(
+//          text: 'Loading...',
+//          textLocation: ProgressUicTextLocation.right,
+//        ),
       ),
     );
   }
 
   Future<List<int>> _getItems(int page) async {
     _loadingAttempts++;
-    await Future.delayed(Duration(seconds: 5), () {});
+    print("_getItems(): page=$page, attemmpt=$_loadingAttempts");
+    await Future.delayed(Duration(seconds: 3));
     List<int> result = List();
     if (_loadingAttempts == 1) {
       return result;
     }
     if (_loadingAttempts == 2) {
       return Future.error('Error loading data');
+    }
+    if (_loadingAttempts == 4) {
+      return result;
+    }
+    if (_loadingAttempts == 6) {
+      return Future.error('Error loading data');
+    }
+    if (page == 11) {
+      return result;
     }
     for (int i = 1; i <= 10; i++) {
       result.add((page - 1) * 10 + i);