+++ lib/list_uic.dart
@@ -12,9 +12,18 @@ import 'progress_uic.dart';
 /// Beside of [controller] you have to define a [itemBuilder] parameter, which
 /// is a function that creates list item widget by the item object.
 ///
-/// ## Pagination
+/// ## Data loading
 ///
-/// *To be implemented in future version.*
+/// **ListUic** supports pull-to-refresh gesture to reload the list items.
+/// While loading, the progress indicator is displayed. If data loading failed,
+/// a snack bar with [errorText] message is shown.
+///
+/// If controller's `allowPagination` property is set to true, loading data page
+/// by page (infinite scrolling) is enabled. When user scroll the list to the end,
+/// the next page of data is loading. The controller manages the current page to load.
+/// During data loading the [nextPageProgressView] widget is displaying at the
+/// end of the list. By default it is a circular progress indicator. You can
+/// provide your custom progress widget for it.
 ///
 /// ## Empty state
 ///
@@ -40,7 +49,8 @@ import 'progress_uic.dart';
 /// * [ListUicController]
 /// * [ListUicEmptyView]
 /// * [ListUicEmptyProgressView]
-/// * [ListView.builder]
+/// * [ListUicNextPageProgressView]
+/// * [ListView]
 ///
 class ListUic<T> extends StatelessWidget {
   ListUic({
@@ -57,16 +67,20 @@ class ListUic<T> extends StatelessWidget {
     Widget emptyErrorView,
     this.emptyProgressText = 'Loading...',
     Widget emptyProgressView,
+    Widget nextPageProgressView,
+    this.errorText = 'Error loading data',
+    this.errorColor = Colors.redAccent,
   }) : assert(emptyDataView != null || emptyDataText != null),
         emptyDataView = emptyDataView ?? ListUicEmptyView(
             controller: controller,
             icon: emptyDataIcon,
             text: emptyDataText),
-        emptyErrorView = emptyDataView ?? ListUicEmptyView(
+        emptyErrorView = emptyErrorView ?? ListUicEmptyView(
             controller: controller,
             icon: emptyErrorIcon,
             text: emptyErrorText),
         emptyProgressView = emptyProgressView ?? ListUicEmptyProgressView(text: emptyProgressText),
+        nextPageProgressView = nextPageProgressView ?? ListUicNextPageProgressView(),
         super(key: key);
 
   /// Manages the list state
@@ -99,6 +113,15 @@ class ListUic<T> extends StatelessWidget {
   /// View to display when the initial data loading is in progress.
   final Widget emptyProgressView;
 
+  /// View to display in the bottom of the list when next page of data is loading.
+  final Widget nextPageProgressView;
+
+  /// Text to display in snack bar when data loading is failed
+  final String errorText;
+
+  /// Color of the error snack bar
+  final Color errorColor;
+
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider.value(
@@ -113,7 +136,10 @@ class ListUic<T> extends StatelessWidget {
             case _ListUicState.emptyError:
               return emptyErrorView;
             case _ListUicState.data:
-              return _dataView();
+            case _ListUicState.progress:
+            case _ListUicState.progressNextPage:
+            case _ListUicState.error:
+              return _buildDataView(context, state.value);
             default:
               return Container();
           }
@@ -122,12 +148,44 @@ class ListUic<T> extends StatelessWidget {
     );
   }
 
-  Widget _dataView() {
-    return ListView.builder(
-      itemCount: controller.items.value.length,
-      itemBuilder: (context, index) {
-        return itemBuilder(controller.items.value[index]);
-      },
+  Widget _buildDataView(BuildContext context, _ListUicState state) {
+    if (state == _ListUicState.error) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        Scaffold.of(context)
+          ..hideCurrentSnackBar()
+          ..showSnackBar(
+            SnackBar(
+              content: Text(errorText),
+              behavior: SnackBarBehavior.floating,
+              backgroundColor: errorColor,
+            )// SnackBar
+          );
+      });
+    }
+    if (state == _ListUicState.progressNextPage) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        controller.scrollController.animateTo(controller.scrollController.position.maxScrollExtent,
+            curve: Curves.linear, duration: Duration (milliseconds: 500));
+      });
+    }
+    int itemCount = (state != _ListUicState.progressNextPage) ?
+        controller.items.value.length :
+        controller.items.value.length + 1;
+    return RefreshIndicator(
+      onRefresh: controller.refresh,
+      child: ListView.builder(
+        physics: const AlwaysScrollableScrollPhysics(),
+        controller: controller.scrollController,
+        itemCount: itemCount,
+        itemBuilder: (context, index) {
+          if (state == _ListUicState.progressNextPage && index == itemCount - 1) {
+            return nextPageProgressView;
+          }
+          else {
+            return itemBuilder(controller.items.value[index]);
+          }
+        },
+      ),
     );
   }
 }
@@ -202,6 +260,27 @@ class ListUicEmptyProgressView extends StatelessWidget {
   }
 }
 
+/// Default progress view for when loading next page of data.
+///
+/// Displays progress indicator at the bottom of the list.
+///
+/// See also:
+/// * [ListUic]
+///
+class ListUicNextPageProgressView extends StatelessWidget {
+  const ListUicNextPageProgressView({
+    Key key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 72.0,
+      child: ProgressUic(),
+    );
+  }
+}
+
 /// A controller for [ListUic] widget.
 ///
 /// It manages list component states and data.
@@ -222,7 +301,16 @@ class ListUicEmptyProgressView extends StatelessWidget {
 /// empty progress view.
 /// - If there are no data, an empty data view is shown.
 /// - If initial data loading failed, an empty error view is shown.
-/// *- Pull to refresh and pagination will be supported in future versions.*
+/// - In normal state, when list items are loaded and shown, pull to refresh
+/// gesture is supported to reload the data. The progress indicator is shown
+/// during the data loading.
+/// - If data loading failed, a snack bar with error message is shown
+/// - When [allowPagination] parameter is set to 'true', the page by page
+/// (infinite scrolling) is enabled. When user scroll the list to the end, the
+/// controller start loading next page of items.
+/// - During loading next page of data, controller forces the list component to
+/// show progress view at the end of the list.
+/// - If next page data loading failed, a snack bar with error message is shown.
 ///
 /// See also:
 /// * [ListUic]
@@ -232,6 +320,7 @@ class ListUicController<T> {
     List<T> items,
     @required this.onGetItems,
     this.initialLoading = true,
+    this.allowPagination = true,
   }) {
     _items = ValueNotifier(items ?? List());
     _page = 1;
@@ -247,6 +336,22 @@ class ListUicController<T> {
     else {
       _state = ValueNotifier(_ListUicState.data);
     }
+
+    if (allowPagination) {
+      _scrollController.addListener(() {
+        if (_scrollController.position.pixels ==
+            _scrollController.position.maxScrollExtent) {
+          if (_readyForNextPage) {
+            _readyForNextPage = false;
+            nextPage();
+          }
+        }
+        else if (_scrollController.position.pixels <=
+            _scrollController.position.maxScrollExtent - 10.0) {
+          _readyForNextPage = true;
+        }
+      });
+    }
   }
 
   ValueNotifier<_ListUicState> _state;
@@ -261,10 +366,20 @@ class ListUicController<T> {
   /// Defaults to 'true'
   bool initialLoading;
 
+  /// Whether pagination is allowed
+  ///
+  /// Defaults to 'true'
+  bool allowPagination;
+
   int _page;
 
   /// Callback to load list items by the page
   Future<List<T>> Function(int) onGetItems;
+
+  ScrollController _scrollController = ScrollController();
+  ScrollController get scrollController => _scrollController;
+
+  bool _readyForNextPage = true;
 
   Future<void> refresh() async {
     // Show progress view
@@ -272,9 +387,12 @@ class ListUicController<T> {
           || _state.value == _ListUicState.emptyError) {
       _state.value = _ListUicState.emptyProgress;
     }
+    else if (_state.value != _ListUicState.emptyProgress){
+      _state.value = _ListUicState.progress;
+    }
     // Load first page of the data
     _page = 1;
-    _loadItems()
+    await _loadItems(_page)
         // Show data
         .then((result) {
           _items.value = result;
@@ -287,22 +405,40 @@ class ListUicController<T> {
         })
         // Show error
         .catchError((error) {
-          _state.value = _ListUicState.emptyError;
+          if (_state.value == _ListUicState.emptyProgress) {
+            _state.value = _ListUicState.emptyError;
+          }
+          else {
+            _state.value = _ListUicState.error;
+          }
         });
   }
 
   Future<void> nextPage() async {
-    _page++;
-    List<T> result = await _loadItems();
-    result.addAll(await onGetItems(_page));
-    items.value = result;
+    if (_state.value == _ListUicState.progressNextPage) {
+      return;
+    }
+    _state.value = _ListUicState.progressNextPage;
+    await _loadItems(_page + 1)
+      .then((result) {
+        if (result.isNotEmpty) {
+          List<T> newItems = _items.value;
+          newItems.addAll(result);
+          _items.value = newItems;
+          _page++;
+        }
+        _state.value = _ListUicState.data;
+      })
+      .catchError((error) {
+        _state.value = _ListUicState.error;
+      });
   }
 
-  Future<List<T>> _loadItems() async {
-    return await onGetItems(_page) ?? List();
+  Future<List<T>> _loadItems(int page) async {
+    print('ListUic::_loadItems()');
+    return await onGetItems(page) ?? List();
   }
 }
 
-enum _ListUicState { init,
-                      emptyData, emptyProgress, emptyError,
-                      data, progress, error }
+enum _ListUicState { emptyData, emptyProgress, emptyError,
+                      data, progress, progressNextPage, error }