+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.2.0] - 2020-03-03
+
+- Support paginatio (infinite scrolling)
+
+## [0.1.0] - 2020-02-26
+
+- Support pull to refresh gesture
+
+## [0.0.3] - 2020-02-26
+
+- Update config
+
+## [0.0.2] - 2020-02-26
+
+- Fix docs
+
 ## [0.0.1] - 2020-02-26
 
 UIC (UI Components) is a set of Flutter widgets that simplifies implementing most used UI cases.
@@ -7,3 +23,4 @@ Each UIC widget uses standard Flutter widgets to implements its view, but adds a
 Here is the list of available components:
 - `ListUic` - Wrapper of ListView, which implements related data loading and state management logic.
 - `ProgressUic` - Wrapper of ProgressIndicator with additional text.
+